+++ packages/core/src/browser.ts
@@ -15,3 +15,4 @@ export * from "./compare.js";
 export * from "./check/index.js";
 export * from "./link.js";
 export * from "./report.js";
+export * from "./ci/github.js";